Allow running SNe binary from any file
Previously, SNe expected to find a parameter file, in.par,
located in the current working directory.  Now it only uses
that path (./in.par) as a default, and allows the path to the
parameter file to be specified as a command line argument.
If the path cannot be found, the program now warns the user
and exits properly, rather than segfaulting without warning.

The executable (bin/SNe) is now compiled from the shared library,
rather than object files.

Also, the shared library was being compiled as an executable,
rather than as a shared library. This has been fixed.