Added restart functionality -- all but initial()
The code is now setup to restart from a previous checkpoint file.
It should set the times, file ids, and checkpoint numbers correctly.

The code is still effectively broken though.
The initial() function currently generates a grid of radii,
and then tries to do a sort of 'previous' interpolation to the
grid that we are trying to restart from.
This is not what we want. We want to be able to restart with
the old grid, rather than trying to resample it.

Fixing the initial() code will take more time,
and possibly require a refactor of the initial conditions
into a proper family of classes.