finished restart capability; style updates
Restart functionality, via an object-oriented approach,
has now been completed.  Now the choice of initial conditions
can be passed in via the parameter file.
This also means that initial conditions WILL NOT be selected
from within the makefile. All initial conditions will need
to be compiled and included (or have the appropriate lines
of the source code stripped out). Overall this should allow it to be
easier to write and develop new initial condition modules.

To do: Hydro (eulerian, etc) and Outputs (ascii, hdf, etc)
could probably benefit from a similar object oriented approach.
Otherwise you'd have to find a hackish way to add the right
include statements within the code.

During this time, I went through and made a lot of style/
whitespace edits.  Sorry that those changes got added into
this commit/branch, but it was much easier to deal with it now,
when I was making a large architecture change.