Add auto-extend functionality
Rather than having to worry if the domain is large enough when we
set the initial conditions, we can instead check to see if the
shock is nearing the outer boundary.  If it is, we add some extra
zones, by simply copying the primitives from the outermost
non-guard cell.

Clearly this will have issues if we want the extended domain
to have non-uniform conditions.  If that's ever needed,
the extend_grid method is virtual and can be overwritten by the
derived class for initial conditions.

At some point, "initial" conditions and the problem/geometry
might be better decoupled.