highlight timestep-limiting cell in plots
In order to better understand what might need to be optomized,
it would be nice to know which cells are keeping the timesteps short.
(In particular, this lets us know if the cells are too refined close in,
or if the calculation is going to be inherently slow.)

By default this is turned off, as it seems like it might be a distractor,
but there's an interactive widget which will allow it to be turned on.