Allow different mass loss prescriptions
The amount of supernova ejecta and the amount of mass ejected
pre-supernova seem to have a significant impact on the
final momentum, modulated by the cooling within the ejecta.

In order to reproducibly show what the choice of mass loss
prescriptions is, I've now included a set of classes
which contain various prescriptions, and which can be
easily extended.