inner boundary condition bug
During my testing I noticed that uniform initial conditions
did not stay uniform. Notably, the innermost cell collapsed,
even when it was slightly overpressured.

The bug was in the source() function.  For the innermost cell,
I need to pass rm=0 to account for there not being pressure
flux through the innermost boundary.  But that means the dV
of the cell is not longer get_dV(rp, rm). This is a problem
since the geometry within source() assumed that dV was valid
for the given rp, rm. While we could use the inaccurate dV
for the hydro source term, we can't use it for calculating
the cooling.

The fix: we can make the geometry of source() more
transparent. This removes the need for a dV in the hydro term,
while also making it more clear what is actually being added.
Hopefully this will help keep things clear in the future.