Fixed restart bugs by clarifying IC classes
The initial conditions model was getting fairly complicated,
since many of the initial conditions classes require certain
computations to be done in a certain order. (E.g. find
a checkpoint to restart from, then set the correct times,
then use that filename to generate an `output_prefix', then
read in the supernovae and filter out the ones which have
already gone off).  It became easy to break this order for one
class when making a change to another.

In order to avoid these problems, I've added some extra
`protected' functions which should only be used in the proper
order.  Also more functionality was added into the IC classes.

Possible change for the future: make setupDomain a method for
each initial condition. Some of the functionality is already
being overwritten by methods like `set_times', so it might be
best to make it an official method, rather than overwriting it
in a piecework way.