now adding some internal energy with the winds
Previously I just added kinetic energy with the winds.
This often wasn't a problem because the winds would shock, and
heat to a reasonable value.

In some cases, this shocking wasn't happening (e.g. in a hot,
diffuse bubble), so temperature remained at a few Kelvin.

So now I'm injecting some thermal energy with the winds.
For simplicity, I'm assuming the wind mass is at 10^4 K.
This internal energy should be much less than the kinetic energy,
but will lead to more reasonable temperatures, pre-shock.