Added better logging of inputs for each run
All the values necessary to re-do a run are now logged in a data
file. This way we won't realize we need access to an input
parameter that we didn't save.  Also, these values should be
automatically parsed as part of Overview. Bonus: these logged
input parameters should also be correctly formatted as a valid
input file.