Added Sedov initial conditions
The true Sedov initial conditions give no ejecta mass;
the explosion is purely energy.  Previous initial conditions
could be hacked to give this setup, but having a true set of
Sedov initial conditions is a little bit nicer.

Note: this still isn't quite the proper Sedov blast wave.
For that to be the case, we would need to have no background
temperature.  Unfortunately, this gets a little ugly, because
we'd need to worry about pressure floors and negative internal
energies.  This might not be too hard, but it's probably
not worth the effort.  As it is, we need not-quite-Sedov
conditions in order to get the canonical stages of SNR evolution.