better attr support for RunSummary objects (i.e. last_run)
The RunSummary objects need to inherit from dictionaries, but
using dict[key] syntax gets ugly and wastes a lot of space,
when I really just want something to behave like an object
attribute.  So I now have `getattr` and `setattr` so that a key
behaves like an attribute, but only if it can't find a real
attribute first.