Simplified Thornton parameter study analysis
Previously, I organized each run in a tree of directories.
Now, to be more consistent with the large cluster parameter study,
I just leave all the Thornton files in one directory.
(I assume that this directory is separate from the cluster
parameter study, but that could be fixed by filtering based on
Inputs.ICs)

Also created a new in.par file for the Thornton runs.