WIP: MPI parallelization
A few of my simulations are taking forever. So I've added some
parallelization back into my code.  Most of this is modelled after
Paul Duffell's incorporation of MPI into RT1D, but it's been tailored
a bit for my code.

THERE ARE STILL BUGS. In particular, thin shells crossing between
processors can lead to some strange behaviors (specifically a sudden
loss of energy and momentum).  It's unclear what exactly is causing
this, but if you restart (allowing the system to re-distribute zones)
then it sometimes fixes the problem.