Bug fix for AMR in restarted runs
The AMR routine needs to have an estimate of the initial zone spacing.
The way it does this is by looking at the *initial number of zones*,
as set in the parameter file.  This usually works fine for runs which
haven't been restarted, since the number of initial zones equals the
number of zones requested in the parameter file.  This breaks down
for restarted runs, where the number of zones is set by the checkpoint
file that you're restarting from; the parameter file value is ignored.

The problem appears when Num_R is significantly different in the
parameter file and the checkpoint file. In that case, the AMR routine
will drive the simulation towards the number of zones set by the
parameter file, which can lead to a sudden burst of zone creation /
destruction. This is bad for a resolution study.

The fix that I'm using is to overwrite the Num_R value that was in
the parameter file with the number of zones in the checkpoint file.

There are still some outstanding potential issues. First, this now
gives a [slightly] different AMR behavior between runs which were
run in one session (which remembered their original Num_R value)
and runs that were restarted midway.  This shouldn't be a huge
difference, but it will be different. I'm not trying to fix that
now, since my routines for restarting jobs don't include any
logic for propogating Num_R values to new parameter files, and it's
not worth trying to add that in now.

Second, there is still a potential issue, in that the "initial" zone
spacing is set by the initial number of zones but the current physical
size of the domain.  This can cause problems if you need to extend the
domain multiple times, as it will tend to reduce your resolution.
One potential fix would be to either use the current number of zones
(which might lead to runaway AMR problems) or use the initial physical
domain size (which has the same discrepancy of restarted/non-restarted
runs discussed in the previous paragraph).  Right now I am leaving that
unchanged, since most of my runs don't need to be extended significantly.