Fixed AMR bug; various small changes
Previously if I encountered an AMR-merge situation with a multiphase
cell, I just added the respective conservative variables together
between each cell.  This isn't truly accurate -- among other things
it doesn't ensure that both phases have the same velocity at the end
of the merge.  I could enforce that by re-writing the momenta, but
then I should also re-write the energies (otherwise the energies
can go negative).  After working at this for a while, I concluded
that there's no unique way to re-write the energies correctly.

So I switched to a different AMR approach.  Choose the left cell
as the primary cell, and merge in the conservatives from the other
[secondary cell] into the primary.  Split the total secondary mass
between the primary's sub-grid phases based on the initial primary
mass fraction of the respective phases.  Split the momentum based
on the mass fraction. Split the metallicity by the metal fraction.
Split the kinetic energy by the mass fraction. Split the internal
energy based on the internal energy fraction. Note that none of
this splitting depends on the multiphase nature of the secondary
cell; all that information is lost. :(

This mirrors what happens to the fluxes + sources, but it has the
downside of artifically choosing a prefered primary cell.
(The fluxes + sources have this degeneracy broken by physics, which
decides which direction the fluxes should go).  This artificial
symmetry breaking might be non-ideal, and there are plenty of
other reasonable alternatives.  Ultimately I chose to preserve the
left cell, since that cell is most likely to have feedback energy
inside it. (If I chose the right cell, it might be difficult for
keller's prescription to propagate outwards, following the shock).
Unfortunately, this means that if the left cell is single phase,
but the right cell is multiphase, the resulting cell will be
single phase.

Separately, I implemented an additional criterion that could trigger
the regression to a cell being single phase: if m_hot > m_cold * 1e5.
In these cases, where the hot mass dominates the cold mass, the cold
energy is very susceptible to rounding errors, which can lead to
negative internal energies. (Usually those problems happen when
m_hot > 1e12 m_cold, near the double precision limit). Note that
I'm not merging if m_cold > 1e5 m_hot; those are precisely the
situations that Keller's method is designed to handle.
This does lead to one potential problem: injecting into a small
cell which then merges with a more massive cell.  In the first step,
the innermost cell might revert back to single-phase, then later it might
merge with a more massive, cold cell. (Due to the r**2 term in the volume,
cell masses grow like r**2, for linearly spaced cells.)  This
might cause my implementation of Keller's method to work less-well
than it should.